[stable @ 92fe97e3b17a12b75498f172e9846dd4a8a973be] compilation warning fixes and other stuff
git-svn-id: svn://svn.code.sf.net/p/colinux/code/branches/stable@153 688f6386-cf32-0410-9ead-a596f0484e40

+++ src/colinux/os/linux/kernel/module/main.c
@@ -32,7 +32,7 @@ static int __init test_module_init(void)
 			printk("colinux: PAE is enabled, cannot continue\n");
 			return -ENOSYS;
 		}
-		printk("colinux: manager load failure: %x\n", rc);
+		printk("colinux: manager load failure: %x\n", (int)rc);
 		return -ENXIO;
 	}
